requirements.txt
Flask
Pillow
python-dotenv
pytest